Fix migration path
